# blog-api/src/main/java/com/example/blog/dto/admin/AdminUserResetPasswordRequest.java
package com.example.blog.dto.admin;

import io.swagger.v3.oas.annotations.media.Schema;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Size;

public class AdminUserResetPasswordRequest {

    @Schema(description = "New password", example = "123456")
    @NotBlank
    @Size(min = 6, max = 100)
    private String newPassword;

    public String getNewPassword() {
        return newPassword;
    }

    public void setNewPassword(String newPassword) {
        this.newPassword = newPassword;
    }
}