# docs/generated/openapi.yaml
openapi: 3.0.1
info:
  title: Blog API
  description: |-
    Auto-generated OpenAPI spec via springdoc-openapi.

    - Web APIs: /api
    - Admin APIs: /api/admin

    Admin APIs require: Authorization: Bearer <token>
  version: v1
servers:
- url: http://localhost:8080
  description: Local dev
security:
- bearer-jwt: []
...
paths:
...
  /api/admin/admins/{id}/status:
    put:
      tags:
      - Admin - Admin Users
      summary: Update admin status
      description: Enable or disable an admin user.
      operationId: status
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserStatusRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
  /api/admin/admins/{id}/reset-password:
    put:
      tags:
      - Admin - Admin Users
      summary: Reset admin password
      operationId: resetPassword
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserResetPasswordRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
...
  /api/admin/admins:
    get:
      tags:
      - Admin - Admin Users
      summary: List admins
      operationId: list_6
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListAdminUserVO'
    post:
      tags:
      - Admin - Admin Users
      summary: Create admin
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMapStringObject'
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: object
          description: Response data payload
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
    AdminUserStatusRequest:
      required:
      - status
      type: object
      properties:
        status:
          maximum: 1
          minimum: 0
          type: integer
          description: "Status: 1=enabled, 0=disabled"
          format: int32
          example: 1
          enum:
          - 0
          - 1
    ApiResponseMapStringObject:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: object
          additionalProperties:
            type: object
            description: Response data payload
          description: Response data payload
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    AdminUserResetPasswordRequest:
      required:
      - newPassword
      type: object
      properties:
        newPassword:
          maxLength: 100
          minLength: 6
          type: string
          description: New password
          example: "123456"
...
    AdminUserCreateRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          type: string
          description: Login username
          example: admin2
        password:
          maxLength: 100
          minLength: 6
          type: string
          description: Initial password
          example: "123456"
        displayName:
          maxLength: 50
          minLength: 0
          type: string
          description: Display name (optional)
          example: Editor
...
    AdminUserVO:
      type: object
      properties:
        id:
          type: integer
          description: Admin id
          format: int64
          example: 1
        username:
          type: string
          description: Username
          example: admin
        displayName:
          type: string
          description: Display name
          example: Administrator
        status:
          type: integer
          description: "Status: 1=enabled, 0=disabled"
          format: int32
          example: 1
          enum:
          - 0
          - 1
        createdAt:
          type: string
          description: Created time
          example: 2026-02-01 12:00:00
        updatedAt:
          type: string
          description: Updated time
          example: 2026-02-01 12:00:00
      description: Admin user
    ApiResponseListAdminUserVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: array
          description: Response data payload
          items:
            $ref: '#/components/schemas/AdminUserVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT